# bid-and-win/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: "BidAndWin API Documentation "
  description: "API for BidAndWin project. This API provides endpoints for Clients,Users,Auctions,Bids,and Items entities."
  contact:
    name: Omer Uruc And Rahim Rzayev
    url: https://github.com/Mruruc
    email: mr.uruc@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/api
    description: Local server for Development.
  - url: https://bid-and-win-web-service-578c86ac472b.herokuapp.com/api
    description: Heroku server for Production.

## tags
...
paths:
...
  /items:
    get:
      tags:
        - Items
      summary: Get all items
      description: Get all items from the database.
      operationId: getItems
      responses:
        "200":
          description: List of all items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemDto"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectInternalServerError"
    post:
      tags:
        - Items
      summary: Create item
      description: Create item in the database.
      operationId: createItem
      requestBody:
        description: Item to be created
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JutsItemDto"
        required: true
      responses:
        "201":
          description: Item created
          content:
            application/url:
              schema:
                type: string
                format: uri
                example: "http://localhost:8080/api/items/1"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/400Error"

        "404":
          description: Client Or Auction Not Foud.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectEntityNotFound"

        "500":
          description: Internal server error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorObject"
  /items/{itemId}:
    get:
      tags:
        - Items
      summary: Get item by id
      description: "Get item by id if exist,else 404 not found ."
      operationId: getItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Item by id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDto"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
    put:
      tags:
        - Items
      summary: Update item by provided item id.
      description: Update item in the database.
      operationId: updateItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
      requestBody:
        description: Item to be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JutsItemDto"
        required: true
      responses:
        "200":
          description: Item updated
          content:
            application/url:
              schema:
                type: string
                format: uri
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/400Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectInternalServerError"
    delete:
      tags:
        - Items
      summary: Delete item by provided id
      description: Delete item in the database.
      operationId: deleteItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "204":
          description: Item deleted.
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectEntityNotFound"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectInternalServerError"
  /items/info/{itemId}:
    get:
      tags:
        - Items
      summary: Download item information by provided item id
      description: Download item information from the database.
      operationId: downloadItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Item downloaded
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                example: /api/items/info/1
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectEntityNotFound"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectInternalServerError"
  /items/client/{clientId}:
    get:
      tags:
        - Items
      summary: Return List of item by client id
      description: "Get items by client id if exist,else 404 not found ."
      operationId: getItems_1
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: Items by client id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemDto"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectEntityNotFound"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectInternalServerError"
  /users/{clientId}:
    post:
      tags:
        - Users
      operationId: createUser
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/UserDTO"
  /users/login:
    post:
      tags:
        - Users
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Return user .
      description: retuen user for we application operation id
      operationId: getUserByClientId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
  /bids/{auctionId}:
    post:
      tags:
        - Bids
      summary: create bids
      description: create bids with specified auctions id
      operationId: saveBid
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidDto"
        required: true
      responses:
        "200":
          description: CREATED
          content:
            application/uri:
              schema:
                type: string
                format: uri
                example: "/api/auctions/1"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/400Error"

        "404":
          description: Client Or Item Not Foud.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectEntityNotFound"

        "500":
          description: Internal server error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorObject"
  /bids/item/{itemId}:
    get:
      tags:
        - Bids
      summary: return list of bid.
      description: return list of bid by provided item id.
      operationId: getBidsByItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BidDto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/400Error"

        "404":
          description: Item Not Foud.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObjectEntityNotFound"

        "500":
          description: Internal server error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorObject"
...
components:
...
  schemas:
...
    BidDto:
      type: object
      properties:
        bidId:
          type: integer
          format: int64
          example: 1
        bidAmount:
          type: number
          format: double
          example: 88.99
        bidTime:
          type: string
          format: date-time
        itemId:
          type: integer
          format: int64
        auctionId:
          type: integer
          format: int64
        clientId:
          type: integer
          format: int64

    ## itemDto
    ItemDto:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
          example: 1
        itemTitle:
          type: string
          example: "Vintage Vase"
        itemDescription:
          type: string
          example: "A beautifully crafted vintage vase from the 19th century."
        startingPrice:
          type: number
          format: double
          example: 50.99
        currentPrice:
          type: number
          format: double
          example: 50.99
        auctionId:
          type: integer
          format: int64
          example: 1
        clientId:
          type: integer
          format: int64
          example: 1
        bids:
          type: array
          items:
            $ref: "#/components/schemas/BidDto"

      ## userDto schema
    UserDTO:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        userName:
          type: string
          example: john@doe
        password:
          type: string
          example: "$2a$10$89CqcTqK5O0toffDbOtnD.fLRZQ/TciLUqY9HP2E.tB4f/1wakf1W"
        failedLoginAttempts:
          type: integer
          format: int64
          example: 0
        lastModified:
          type: string
          format: date
          example: "2024-05-23"
        clientId:
          type: integer
          format: int64
          example: 1
        accountLocked:
          type: boolean
          example: false

    ## auctionDto schema
...
    JutsItemDto:
      type: object
      properties:
        itemTitle:
          type: string
          example: "Vintage Vase"
        itemDescription:
          type: string
          example: "A beautifully crafted vintage vase from the 19th century."
        startingPrice:
          type: number
          format: double
          example: 50.99
        currentPrice:
          type: number
          format: double
          example: 50.99
        auctionId:
          type: integer
          format: int64
          example: 1
        clientId:
          type: integer
          format: int64
          example: 1

    ## error objects
    ErrorObject:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorObjectEntityNotFound:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
          example: Entity Not Found!
        timestamp:
          type: string
          format: date-time
          example: 2021-06-01T00:00:00
    ErrorObjectInternalServerError:
      type: object
      properties:
        status:
          type: integer
          example: 500
        message:
          type: string
          example: Internal Server Error!
        timestamp:
          type: string
          format: date-time
          example: 2021-06-01T00:00:00
...
    400Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid Field/s!
        timestamp:
          type: string
          format: date-time
          example: 2021-06-01T00:00:00