# bid-and-win/web-service/src/main/java/com/bidAndWin/controller/UserController.java
package com.bidAndWin.controller;

import com.bidAndWin.dto.UserDTO;
import com.bidAndWin.service.UserService;
import com.fasterxml.jackson.core.JsonProcessingException;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@CrossOrigin(origins = "*")
@RestController
@RequestMapping("/users")
@RequiredArgsConstructor

public class UserController {
    private final UserService userService;

    @GetMapping(value = "/{userId}", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<UserDTO> getUserByClientId(@PathVariable Long userId) {
        UserDTO userDto = userService.getUserById(userId);
        return new ResponseEntity<>(userDto, HttpStatus.OK);
    }

    @PostMapping(value = "/{clientId}", consumes = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<UserDTO> createUser(@PathVariable Long clientId, @RequestBody UserDTO userDTO) throws JsonProcessingException {
        return new ResponseEntity<>(userService.saveUser(clientId,userDTO), HttpStatus.CREATED);
    }

    @PostMapping(value = "/login", consumes = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<UserDTO> loginUser(@RequestBody UserDTO userDto) {
        return new ResponseEntity<>(userService.login(userDto), HttpStatus.OK);
    }
}